requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
kafka-python
sqlalchemy
pydantic

